Refactored Car entity and CarDTO. Wrote integration test and unit for CarService

+++ src/main/java/com/rental/CarRentalShop/dto/CarDTO.java
@@ -3,6 +3,8 @@ package com.rental.CarRentalShop.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class CarDTO {
@@ -11,5 +13,5 @@ public class CarDTO {
     private String model;
     private Integer year;
     private String registrationNumber;
-    private Double rentalPrice;
+    private BigDecimal rentalPrice;
 }
